portpholio/requirements.txt
python-multipart
yfinance
numpy
scikit-learn
requests
pydantic
openpyxl
duckduckgo-search
phidata